#10 [REFACTOR] : 패키지 리팩토링
