+++ app/src/main/java/co/kr/soptandroidseminar/data/main/profile/RepoData.kt
@@ -1,4 +1,4 @@
-package co.kr.soptandroidseminar.data
+package co.kr.soptandroidseminar.data.main.profile
 
 data class RepoData(
     val name: String?,
