+++ app/src/main/java/co/kr/soptandroidseminar/data/main/profile/ResponseFollowerData.kt
@@ -1,4 +1,4 @@
-package co.kr.soptandroidseminar.data
+package co.kr.soptandroidseminar.data.main.profile
 
 data class ResponseFollowerData(
     val login: String,
