+++ app/src/main/java/co/kr/soptandroidseminar/data/main/profile/ResponseUserInfoData.kt
@@ -1,4 +1,4 @@
-package co.kr.soptandroidseminar.data
+package co.kr.soptandroidseminar.data.main.profile
 
 data class ResponseUserInfoData(
     val avatar_url: String,
