+++ app/src/main/java/co/kr/soptandroidseminar/data/signin/RequestSignInData.kt
@@ -1,4 +1,4 @@
-package co.kr.soptandroidseminar.data
+package co.kr.soptandroidseminar.data.signin
 
 data class RequestSignInData(
     val email: String,
