+++ app/src/main/java/co/kr/soptandroidseminar/data/signin/ResponseSignInData.kt
@@ -1,15 +1,14 @@
-package co.kr.soptandroidseminar.data
+package co.kr.soptandroidseminar.data.signin
 
-data class ResponseSignUpData(
+data class ResponseSignInData(
     val status: Int,
     val success: Boolean,
     val message: String,
     val data: Data
-) {
+)  {
     data class Data(
         val id: Int,
         val name: String,
-        val password: String,
         val email: String,
     )
 }
