+++ app/src/main/java/co/kr/soptandroidseminar/data/signup/ResponseSignUpData.kt
@@ -1,14 +1,15 @@
-package co.kr.soptandroidseminar.data
+package co.kr.soptandroidseminar.data.signup
 
-data class ResponseSignInData(
+data class ResponseSignUpData(
     val status: Int,
     val success: Boolean,
     val message: String,
     val data: Data
-)  {
+) {
     data class Data(
         val id: Int,
         val name: String,
+        val password: String,
         val email: String,
     )
 }
