+++ app/src/main/res/layout/fragment_follower.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.profile.FollowerFragment">
+    tools:context=".view.main.profile.FollowerFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rcv_follower"
