+++ app/src/main/res/layout/fragment_home_follower.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/empty_fragment"
-    tools:context=".view.home.HomeFollowerFragment">
+    tools:context=".view.main.home.HomeFollowerFragment">
 
     <TextView
         android:layout_width="match_parent"
