+++ app/src/main/res/layout/fragment_home_following.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/empty_fragment"
-    tools:context=".view.home.HomeFollowingFragment">
+    tools:context=".view.main.home.HomeFollowingFragment">
 
     <TextView
         android:layout_width="match_parent"
