+++ app/src/main/res/layout/fragment_repo.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.profile.RepoFragment">
+    tools:context=".view.main.profile.RepoFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rcv_repo"
